requirements.txt
azure-functions
requests
pandas
azure-storage-blob